EAMxx: Move read utils files to readfiles folder.
